added license/copyright headers, fixed indentation of preprocessor directives
