+++ src/main/main.pri
@@ -1,6 +1,6 @@
 macx {
-    OBJECTIVE_SOURCES += $$PWD/MacOsEventListener.mm
-    HEADERS += $$PWD/MacOsEventListener.h
+    OBJECTIVE_SOURCES += $$PWD/mac/MacOsEventListener.mm
+    HEADERS += $$PWD/mac/MacOsEventListener.h
     LIBS += -framework AppKit -framework Foundation
 
 }
